refactor: Heading1 컴포넌트 추출
