+++ src/pages/LinePage/LinePage.style.ts
@@ -17,14 +17,6 @@ const Icon = styled.span`
   }
 `;
 
-const Heading1 = styled.h1`
-  text-align: center;
-  font-size: 1.5rem;
-  font-weight: 600;
-  color: ${PALETTE.GRAY_600};
-  margin-bottom: 1rem;
-`;
-
 const TitleBox = styled<any>(Box)`
   & button[type='button'] {
     position: absolute;
@@ -117,14 +109,4 @@ const StationSelectError = styled.div`
   color: ${PALETTE.RED};
 `;
 
-export {
-  Container,
-  Heading1,
-  Icon,
-  TitleBox,
-  FormBox,
-  Form,
-  List,
-  StationSelects,
-  StationSelectError,
-};
+export { Container, Icon, TitleBox, FormBox, Form, List, StationSelects, StationSelectError };
